requirements.txt
flask
openai
python-dotenv
numpy